requirements.txt
amqp==2.2.1
anyjson==0.3.3
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.3
cairocffi==0.8.0
CairoSVG==1.0.22
celery==4.1.0
certifi==2017.1.23
cffi==1.10.0
cssselect2==0.2.0
Django==1.9.9
django-admin-tools==0.8.1
django-allauth==0.28.0
django-ckeditor==5.1.1
django-embed-video==1.1.0
django-mptt==0.8.6
django-paypal==0.3.6
django-rest-auth==0.8.2
django-tinymce==2.4.0
djangorestframework==3.5.3
flower==0.9.1
futures==3.0.5
html5lib==0.999999999
kombu==4.1.0
mysqlclient==1.3.9
oauthlib==2.0.0
Pillow==3.4.2
pkg-resources==0.0.0
pycparser==2.18
Pyphen==0.9.4
python-openid==2.2.5
pytz==2016.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
tinycss2==0.6.0
tornado==4.2
vine==1.1.4
WeasyPrint==0.40
webencodings==0.5.1
